Feature/windows no run (#604)
* Just let windows through, seeing what happens

* Add RDP on build finish

* Try not mirroring stdout

* Add windows based polling

* Add stdout / stderr mirroring.  Move platform check to platform.system.  Allow sweeps

* Conditional RDP in appveyor, make output.log work

* Flush or log file in the windows user process

* Proper redirecting in windows

* New attempt at stdout/err logging

* Another windows stdout/err attempt

* Try passing stdout/err

* A WILD attempt at getting PIPEs working

* Lets try this with just writing to files instead of the pipe

* Got windows into a kinda working state

* Fix py2 flake8 errors

* No need for flush, fix flaky spec

* Fix syntaz

+++ wandb/run_manager.py
@@ -341,7 +341,6 @@ class FileEventHandlerTextStream(FileEventHandler):
             self._tailer.stop()
             self._tailer = None
 
-
 class FileEventHandlerBinaryStream(FileEventHandler):
     def __init__(self, *args, **kwargs):
         super(FileEventHandlerBinaryStream, self).__init__(*args, **kwargs)
@@ -401,7 +400,11 @@ class Process(object):
     def poll(self):
         if self.returncode is None:
             try:
-                os.kill(self.pid, 0)
+                if platform.system() == "Windows":
+                    if windows.pid_running(self.pid) == False:
+                        raise OSError(0, "Process isn't running")
+                else:
+                    os.kill(self.pid, 0)
             except OSError as err:
                 if err.errno == errno.ESRCH:
                     # ESRCH == No such process
@@ -1058,7 +1061,7 @@ class RunManager(object):
         """
         stdout_streams, stderr_streams = self._get_stdout_stderr_streams()
 
-        if sys.platform == "win32":
+        if platform.system() == "Windows":
             # PTYs don't work in windows so we use pipes.
             self._stdout_tee = io_wrap.Tee.pipe(*stdout_streams)
             self._stderr_tee = io_wrap.Tee.pipe(*stderr_streams)
@@ -1075,7 +1078,7 @@ class RunManager(object):
         runner = util.find_runner(program)
         if runner:
             command = runner + command
-        if sys.platform == "win32":
+        if platform.system() == "Windows":
             command = ' '.join(windows.quote_arg(arg) for arg in command)
         else:
             command = ' '.join(six.moves.shlex_quote(arg) for arg in command)
